% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{get_article_url}
\alias{get_article_url}
\title{Scraps an OJS issue page for articles urls}
\usage{
get_article_url(
  input_url,
  use_conventional_url = TRUE,
  verbose = FALSE,
  method = "scrap_by_href_convention_no_classes"
)
}
\arguments{
\item{use_conventional_url}{Logical. Should ojsr parse the urls given to form the conventional url to scrap?}

\item{verbose}{Logical.}

\item{method}{String. Available methods: scrap_by_href_convention_no_classes, scrap_by_href_convention_no}

\item{url}{Character vector.}
}
\value{
A dataframe with the urls of the articles linked from the OJS issue page.
}
\description{
This functions xxx
}
\details{
(Warning: This will only work on OJS v1,v2 installations with none or min customization. Please refer to vignette.)
}
\examples{

    xxx

}
