% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{get_article_urls_from_issue}
\alias{get_article_urls_from_issue}
\title{Scraps an OJS issue page for articles urls}
\usage{
get_article_urls_from_issue(
  url,
  method = "scrap_by_href_convention_no_classes",
  verbose = FALSE
)
}
\arguments{
\item{url}{Character vector.}

\item{method}{String. Available methods: scrap_by_href_convention_no_classes, scrap_by_href_convention_no}

\item{verbose}{Logical.}
}
\value{
A dataframe with the urls of the articles linked from the OJS issue page.
}
\description{
This functions xxx
}
\details{
(Warning: This will only work on OJS v1,v2 installations with none or min customization. Please refer to vignette.)
}
\examples{

    xxx

}
