% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{get_articles_urls_from_issue}
\alias{get_articles_urls_from_issue}
\title{Scraps an OJS issue page for articles urls}
\usage{
get_articles_urls_from_issue(url, verbose = FALSE)
}
\arguments{
\item{url}{Character vector.}

\item{verbose}{Logical.}
}
\value{
A dataframe with the urls of the articles linked from the OJS issue page.
}
\description{
This functions takes an array of OJS issues URLs and scrap the articles urls from their links.
}
\details{
Before scraping, URL is checked to comply to an OJS issue URL convention.
Then, this function retrieves all the hrefs from links with usual OJS classes (e.g., a.tocTitle)

(Warning: This will only work on OJS v1,v2 installations with none or min customization. Please refer to vignette.)
}
\examples{
get_articles_urls_from_issue(c('https://publicaciones.sociales.uba.ar/index.php/psicologiasocial/issue/view/319/showToc',
    'https://firstmonday.org/ojs/index.php/fm/issue/view/634'))
}
