% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{get_galley_urls}
\alias{get_galley_urls}
\title{Scraps the galley(s) url(s) from article/abstract OJS pages}
\usage{
get_galley_urls(url, verbose = FALSE)
}
\arguments{
\item{url}{Character vector.}

\item{verbose}{Logical.}
}
\value{
A dataframe with the urls of the galleys linked from the OJS article page. (Please refer to vignette for structure)
}
\description{
This functions takes an array of OJS article/abstract URLs and scrap the galley(s) url from their links.
}
\details{
"Galleys" are the public (and usually full-content) copies of the articles. Usual formats for galleys are pdf, xml, html, rtf, epub.
These may also include supplementary materials (dataset, charts, etc.).

Before scraping, URL is checked to comply to an article/abstract URL convention.
Then, this function retrieves all the hrefs from links with usual OJS classes (e.g., #obj_galley_link)

(Warning: This will only work on OJS v1,v2 installations with none or min customization. Please refer to vignette.)
}
\examples{
get_galley_urls(c('https://firstmonday.org/ojs/index.php/fm/article/view/9540',
    'http://imed.pub/ojs/index.php/iam/article/view/1650'))
}
