% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{get_galleys_from_article}
\alias{get_galleys_from_article}
\title{Scraping galleys URLs from OJS articles}
\usage{
get_galleys_from_article(input_url, verbose = FALSE)
}
\arguments{
\item{input_url}{Character vector.}

\item{verbose}{Logical.}
}
\value{
A long-format dataframe with the url you provided (input_url), the articles url scrapped (output_url),
the format of the galley (format), and the url that forces download of the galley (download_url)
}
\description{
Takes a vector of OJS URLs and scrapes all the galleys URLs from the article view
}
\examples{
\donttest{
article <- 'https://revistapsicologia.uchile.cl/index.php/RDP/article/view/55657'
galleys <- ojsr::get_galleys_from_article(input_url = article)
}
}
