% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{get_issue_urls_from_archive}
\alias{get_issue_urls_from_archive}
\title{Scraps an OJS archive page for issues urls}
\usage{
get_issue_urls_from_archive(url, verbose = FALSE)
}
\arguments{
\item{url}{Character vector.}

\item{verbose}{Logical.}
}
\value{
A dataframe with the urls of the articles linked from the OJS issue page.
}
\description{
This functions takes an array of OJS archive URLs and scrap the issue urls from their links.
}
\details{
Before scraping, URL is checked to comply to an OJS archive URL convention.
Then, this function retrieves all the hrefs from links with usual OJS classes (e.g., a.tocTitle)

(Warning: This will only work on OJS v1,v2 installations with none or min customization. Please refer to vignette.)
}
\examples{
get_articles_urls_from_issue(c('https://publicaciones.sociales.uba.ar/index.php/psicologiasocial/issue/view/319/showToc',
    'https://firstmonday.org/ojs/index.php/fm/issue/view/634'))
}
