% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{get_meta_from_html}
\alias{get_meta_from_html}
\title{Scraps metadata from html of OJS pages}
\usage{
get_meta_from_html(
  input_url,
  use_conventional_url = TRUE,
  method = "scrap_meta_in_head",
  verbose = FALSE
)
}
\arguments{
\item{input_url}{Character vector.}

\item{use_conventional_url}{Logical. Should ojsr parse the urls given to form the conventional url to scrap?}

\item{method}{String. Available methods: scrap_meta_in_head}

\item{verbose}{Logical.}
}
\value{
A long-format dataframe with the url you provided (input_url), the name of the metadata (meta_data_name),
the content of the metadata (meta_data_content), the standard in which the content is annotated (meta_data_scheme),
and the language in which the metadata was entered (meta_data_xmllang)
}
\description{
Takes a vector of OJS urls and and scraps the metadata written in the html.
}
\details{
Search criteria: <meta> tags in the <head> section of the html (method='scrap_meta_in_head')
}
\examples{

    socPsy_articles <- c(
        'https://publicaciones.sociales.uba.ar/index.php/psicologiasocial/article/view/2137',
        'https://dspace.palermo.edu/ojs/index.php/psicodebate/article/view/516/311'
    )
    metadata <- ojsr::get_meta_from_html(socPsy_articles,
        use_conventional_url = TRUE, verbose = TRUE)

}
