% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{get_oai_meta_from_article}
\alias{get_oai_meta_from_article}
\title{Get OAI metadata from an OJS article url}
\usage{
get_oai_meta_from_article(input_url, verbose = FALSE)
}
\arguments{
\item{input_url}{Character vector.}

\item{verbose}{Logical.}
}
\value{
A long-format dataframe with the url you provided (input_url), the name of the metadata (meta_data_name),
and the content of the metadata (meta_data_content).
}
\description{
This functions access OAI records (within OJS) for any article for which you provided an url.
}
\details{
Several limitations are in place. Please refer to vignette.
}
\examples{
\donttest{
articles <- c(
  'https://publicaciones.sociales.uba.ar/index.php/psicologiasocial/article/view/2137', # article
  'https://dspace.palermo.edu/ojs/index.php/psicodebate/article/view/516/311' # xml galley
)
metadata_oai <- ojsr::get_oai_meta_from_article(input_url = articles, verbose = TRUE)
}
}
