% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{get_oai_metadata_from_article}
\alias{get_oai_metadata_from_article}
\title{Get OAI metadata for an article}
\usage{
get_oai_metadata_from_article(url, verbose = FALSE)
}
\arguments{
\item{url}{Character vector.}

\item{verbose}{Logical.}
}
\value{
A dataframe with the urls of the articles linked from the OJS issue page.
}
\description{
This functions xxx
}
\details{
(Warning: This will only work on OJS v1,v2 installations with none or min customization. Please refer to vignette.)
}
\examples{

    xxx

}
