% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{get_paginated_search_url}
\alias{get_paginated_search_url}
\title{Creates OJS search result urls}
\usage{
get_paginated_search_url(input_url, search_criteria, verbose = FALSE)
}
\arguments{
\item{input_url}{Character vector.}

\item{search_criteria}{Character string}

\item{verbose}{Logical.}

\item{check_pagination}{Logical.}
}
\value{
A dataframe with the urls of the articles linked from the OJS issue page.
}
\description{
Takes a vector of OJS urls and a string for search criteria to compose the search url.
}
\details{
You may then pass these urls to any other get_*_url() function.

If check_pagination = TRUE, it runs the search in OJS to see if pagination is involved,
and returns the url for the search result pages.
}
\examples{

    socPsy_journals <- c( # 2 social psychology journals
        'https://revistapsicologia.uchile.cl/index.php/RDP/',
        'https://publicaciones.sociales.uba.ar/index.php/psicologiasocial/issue/current'
    )
    search_criteria <- "social representations"
    search_results <- ojsr::get_search_url(socPsy_journals,
        search_criteria = "psicología social", check_pagination = TRUE, verbose = TRUE)

}
