% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{metadata_list_to_df}
\alias{metadata_list_to_df}
\title{Converts a list of scrapped meta-data to a dataframe}
\usage{
metadata_list_to_df(metadataList, url)
}
\arguments{
\item{metadataList}{A list generated with ojsr::get_metadata_from_page}

\item{url}{Character vector (same used in ojsr::get_metadata_from_page).}
}
\value{
A dataframe
}
\description{
Converts a list of scrapped meta-data to a dataframe
}
\examples{
article_URL <- c('https://firstmonday.org/ojs/index.php/fm/article/view/9540',
    'http://imed.pub/ojs/index.php/iam/article/view/1650')
metadata_articles <- get_metadata_from_page( article_URL )
metadata_articles_df <- metadata_list_to_df( metadata_articles , url )
}
