% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojsr-package.R
\docType{package}
\name{ojsr-package}
\alias{ojsr}
\alias{ojsr-package}
\title{ojsr: Crawler and Data Scraper for Open Journal System ('OJS')}
\description{
Crawler for 'OJS' pages and scraper for meta-data from articles. You can crawl 'OJS' archives, issues, articles, galleys, and search results. You can scrape articles metadata from their head tag in html, or from Open Archives Initiative ('OAI') records. Most of these functions rely on 'OJS' routing conventions (\url{https://docs.pkp.sfu.ca/dev/documentation/en/architecture-routes}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gastonbecerra/ojsr}
  \item Report bugs at \url{https://github.com/gastonbecerra/ojsr/issues}
}

}
\author{
\strong{Maintainer}: Gaston Becerra \email{gaston.becerra@gmail.com} (\href{https://orcid.org/0000-0001-9432-8848}{ORCID})

}
\keyword{internal}
