% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_parsers.R
\name{process_URL}
\alias{process_URL}
\title{Parses URLs according to OJS (v1,v2) conventions}
\usage{
process_URL(url)
}
\arguments{
\item{url}{String vector.}
}
\value{
A dataframe wich indicates which type of OJS page the URL refers to:
    "article view" usually has the article's abstract, the full-content links and the references;
    "issue" usually has the table of contents (linking to article views);
    "article galley" usually allows you to download or read the article's full-content;
    "oai" indicates that the URL points to the OAI interface.
    If "article view" or "article galley", the corresponding IDs are returned.
}
\description{
Processes URLs to check which (if any) OJS view it is, if there are IDs for
articles and/or galleys, and the (expected) OAI address.
}
\details{
This fn works by parsing URLs strings against OJS routing conventions.
Of course, this will only work on OJS installations with none or min customization.
It does not check anything against the actual pages!
}
\examples{
process_URL(c('https://firstmonday.org/ojs/index.php/fm/article/view/9540',
    'http://imed.pub/ojs/index.php/iam/article/view/1650'))
}
