% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{get_article_url}
\alias{get_article_url}
\title{Scraps an OJS page for article links}
\usage{
get_article_url(
  input_url,
  use_conventional_url = TRUE,
  method = "scrap_by_href_convention_no_classes",
  verbose = FALSE
)
}
\arguments{
\item{input_url}{Character vector.}

\item{use_conventional_url}{Logical. Should ojsr parse the urls given to form the conventional url to scrap?}

\item{method}{String. Available methods: scrap_by_href_convention_no_classes (default), scrap_by_href_convention}

\item{verbose}{Logical.}
}
\value{
A long-format dataframe with the url you provided (input_url) and the articles url scraped (output_url)
}
\description{
Takes a vector of OJS urls and scraps them to retrieve the links to OJS articles
}
\details{
Search criteria: links containing "/article/view” (method='scrap_by_href_convention_no_classes', default), and the same without filtering (method='scrap_by_href_convention'). (Please refer to vignette.)

(Warning: This will only work on OJS v1,v2 installations with none or min customization. Please refer to vignette.)
}
