% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{get_galley_urls}
\alias{get_galley_urls}
\title{Scraps the galley(s) url(s) from article_abstract OJS pages}
\usage{
get_galley_urls(url, verbose = FALSE)
}
\arguments{
\item{url}{Character vector.}

\item{verbose}{Logical.}
}
\value{
A dataframe with the urls of the galleys linked from the OJS article page. (Please refer to vignette for structure)
}
\description{
This functions takes an array of OJS urls pointing to articles pages and scrap the galley(s) url from their links.
}
\details{
(Warning: This will only work on OJS v1,v2 installations with none or min customization. Please refer to vignette.)
}
\examples{
get_galley_urls(c('https://firstmonday.org/ojs/index.php/fm/article/view/9540',
    'http://imed.pub/ojs/index.php/iam/article/view/1650'))
}
