% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{get_issue_urls_from_archive}
\alias{get_issue_urls_from_archive}
\title{Scraps an OJS archive page for issues urls}
\usage{
get_issue_urls_from_archive(
  url,
  use_conventional_url = TRUE,
  verbose = FALSE,
  method = "scrap_by_href_convention"
)
}
\arguments{
\item{url}{Character vector.}

\item{use_conventional_url}{Logical. Should ojsr parse the urls given to form the conventional url to scrap?}

\item{verbose}{Logical.}

\item{method}{String. Available methods: scrap_by_href_convention}
}
\value{
A dataframe with the urls of the articles linked from the OJS issue page.
}
\description{
This functions xxx
}
\details{
(Warning: This will only work on OJS v1,v2 installations with none or min customization. Please refer to vignette.)
}
\examples{

    xxx

}
