% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{get_issues_from_archive}
\alias{get_issues_from_archive}
\title{Scraping issues’ URLs from the OJS issues archive}
\usage{
get_issues_from_archive(input_url, verbose = FALSE)
}
\arguments{
\item{input_url}{Character vector.}

\item{verbose}{Logical.}
}
\value{
A long-format dataframe with the url you provided (input_url) and the url of issues found (output_url)
}
\description{
Takes a vector of OJS URLs and scrapes the issues URLs from the issue archive.
}
\examples{
\donttest{
journal <- 'https://dspace.palermo.edu/ojs/index.php/psicodebate/issue/archive'
issues <- ojsr::get_issues_from_archive(input_url = journal)
}

}
