% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{get_metadata_from_article}
\alias{get_metadata_from_article}
\title{Scraps OJS article pages for galleys urls}
\usage{
get_metadata_from_article(url, method = "scrap_meta_in_head", verbose = FALSE)
}
\arguments{
\item{url}{Character vector.}

\item{method}{String. Available methods: scrap_by_class_convention}

\item{verbose}{Logical.}
}
\value{
A dataframe with the urls of the articles linked from the OJS issue page.
}
\description{
This functions xxx
}
\details{
(Warning: This will only work on OJS v1,v2 installations with none or min customization. Please refer to vignette.)
}
\examples{

    xxx

}
