% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{get_search_url}
\alias{get_search_url}
\title{xxxx}
\usage{
get_search_url(
  input_url,
  search_criteria,
  check_pagination = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{Logical.}

\item{url}{Character vector.}

\item{use_conventional_url}{Logical. Should ojsr parse the urls given to form the conventional url to scrap?}

\item{method}{String. Available methods: scrap_by_href_convention}
}
\value{
A dataframe with the urls of the articles linked from the OJS issue page.
}
\description{
This functions xxx
}
\details{
(Warning: This will only work on OJS v1,v2 installations with none or min customization. Please refer to vignette.)
}
\examples{

    xxx

}
