% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojsr.R
\docType{package}
\name{ojsr}
\alias{ojsr}
\title{ojsr: A package for scraping OJS}
\description{
This package allows you to crawl OJS archives, issues, articles, galleys,
and search results, and retrieve metadata from articles.
This is useful if OJS Rest API is not available (as in installations prior to v3.1).
It also includes function to parse OJS specific URL conventions.
}
