% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojsr.R
\docType{package}
\name{ojsr}
\alias{ojsr}
\title{ojsr: A package for scrapping OJS}
\description{
This package allows you scrap content (bibliographic metadata) from OJS front-pages and their OAI interfaces;
This is useful when the OJS Rest API is not available (as in OJS installments prior to v3.1).
It also includes function to parse OJS specific URL conventions.
}
