% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_parsers.R
\name{process_urls}
\alias{process_urls}
\title{Parses URLs according to OJS conventions}
\usage{
process_urls(url)
}
\arguments{
\item{url}{Character vector.}
}
\value{
A dataframe wich indicates which type of OJS page the url refers to. Please refer to vignette ("Parse OJS urls" section).
}
\description{
This function parse a string vector of URLs to check which (if any) OJS view they could be referring to,
if there are IDs for articles and/or galleys,
and the (expected) OAI address.
}
\details{
This works by parsing URLs strings against OJS routing conventions. It does not check anything against the actual pages!

(Warning: This will only work on OJS v1,v2 installations with none or min customization. Please refer to vignette.)
}
\examples{
process_urls(c('https://firstmonday.org/ojs/index.php/fm/article/view/9540',
    'http://imed.pub/ojs/index.php/iam/article/view/1650'))
}
