% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_parsers.R
\name{process_urls}
\alias{process_urls}
\title{Parses URLs according to OJS conventions}
\usage{
process_urls(url)
}
\arguments{
\item{url}{Character vector of OJS url(s).}
}
\value{
A dataframe wich indicates the type of OJS url expected, and the parameters.
Please refer to vignette.
}
\description{
This function parse a string vector of OJS urls to extract IDs and generate expected ulrs for the scraping functions.
}
\details{
This works by parsing URLs strings against OJS routing conventions. It does not check anything against the actual pages!
}
\examples{
process_urls(c('https://firstmonday.org/ojs/index.php/fm/article/view/9540',
    'http://imed.pub/ojs/index.php/iam/article/view/1650'))
}
